added a warmup executable that only warms up and save to file

+++ cpp/source/xy/xyHelper.h
@@ -2,8 +2,8 @@
 #define XYHELPER_H
 
 struct WNandNS{
-	WNandNS(int winding_number, long double num_steps) : winding_number(winding_number), num_steps(num_steps) {};
-	int winding_number;
+	WNandNS(double winding_number, long double num_steps) : winding_number(winding_number), num_steps(num_steps) {};
+	double winding_number;
 	long double num_steps;
 };
